requirements.txt
certifi==2017.4.17
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.2
django-allauth==0.32.0
django-ckeditor==5.3.0
django-crispy-forms==1.6.1
django-filter==1.0.4
django-js-asset==0.1.1
djangorestframework==3.6.3
idna==2.5
Markdown==2.6.9
oauthlib==2.0.2
olefile==0.44
Pillow==4.1.1
python3-openid==3.1.0
pytz==2017.2
requests==2.18.1
requests-oauthlib==0.8.0
urllib3==1.21.1
